storage: Move primitives to a separate package
