+++ storage/aliases.go
@@ -1,0 +1,37 @@
+// Copyright 2019 Google Inc. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package storage
+
+import "github.com/google/trillian/storage/tree"
+
+// TODO(pavelkalinnikov): These aliases were created to not break the code that
+// depended on these type. We should delete this file eventually.
+
+type NodeID = tree.NodeID
+type Node = tree.Node
+type Suffix = tree.Suffix
+type PopulateSubtreeFunc = tree.PopulateSubtreeFunc
+type PrepareSubtreeWriteFunc = tree.PrepareSubtreeWriteFunc
+
+var (
+	NewNodeIDFromHash         = tree.NewNodeIDFromHash
+	NewNodeIDFromPrefix       = tree.NewNodeIDFromPrefix
+	NewNodeIDFromBigInt       = tree.NewNodeIDFromBigInt
+	NewNodeIDForTreeCoords    = tree.NewNodeIDForTreeCoords
+	NewNodeIDFromPrefixSuffix = tree.NewNodeIDFromPrefixSuffix
+
+	EmptySuffix = tree.EmptySuffix
+	ParseSuffix = tree.ParseSuffix
+)
